// types.ts
export enum UserRole {
  ADMIN = 'ADMIN',
  TEACHER = 'TEACHER',
  PARENT = 'PARENT',
  BOARD_MEMBER = 'BOARD_MEMBER',
}

export enum SchoolStatus {
    ACTIVE = 'ACTIVE',
    CAUTION = 'CAUTION',
    SUSPENDED = 'SUSPENDED'
}

export interface School {
    id: string;
    name: string;
    address: string;
    motto?: string;
    domain: string;
    badgeUrl?: string | null;
    status: SchoolStatus;
    cautionMessage: string;
    activeSubjectIds: string[];
}


export interface User {
  id: string;
  username: string;
  password?: string; // Password stored here for mock auth
  roles: UserRole[];
  name: string;
  phoneNumber?: string; // For SMS notifications
  studentId?: string; // For parents, links to their child
  teacherId?: string; // For teachers
  profileImageUrl?: string; // For teacher profile images
  signatureImageUrl?: string; // For teacher/admin signatures
  classSubjectAssignments?: { classId: string; subjectIds: string[] }[];
  schoolId?: string; // To associate user with a school. Optional for superadmin.
  isBoardMember?: boolean;
}

export type TeacherFormData = Omit<User, 'id' | 'studentId' | 'profileImageUrl' | 'signatureImageUrl' | 'schoolId'> & { id?: string, password_param?: string };
export type UserProfileFormData = Partial<Pick<User, 'name' | 'password'>> & { profileImageFile?: File | null, signatureImageFile?: File | null };

export enum Gender {
    MALE = 'Male',
    FEMALE = 'Female',
}

export interface Student {
  id:string;
  admissionNumber: string;
  name: string;
  classId: string;
  stream: string; // e.g., "A", "B", "North"
  parentUsername: string;
  profileImageUrl?: string; // For student images on reports
  gender: Gender;
  currentFeesBalance: number;
  nextTermFees: number;
  subjects?: string[]; // Array of subject IDs the student takes
  schoolId: string;
}

export type StudentFormData = Omit<Student, 'id' | 'profileImageUrl' | 'schoolId'> & { id?: string, profileImageFile?: File | null };


export interface SchoolClass {
  id: string;
  name: string; // e.g., "Form 1", "Form 2"
}

export interface Subject {
  id: string;
  name: string;
  group?: number; // To categorize subjects (1=Compulsory, 2=Sciences, etc.)
}

export enum TermCalculationMode {
  WEIGHTED_AVERAGE = 'WEIGHTED_AVERAGE',
  SIMPLE_AVERAGE = 'SIMPLE_AVERAGE',
}

export interface Term {
  id: string;
  name: string; // e.g., "Term 1", "Term 2"
  year: number;
  calculationMode: TermCalculationMode;
  closingDate?: string;
  openingDate?: string;
  schoolId: string;
}

export interface ExamSession {
  id: string;
  termId: string;
  name: string; // e.g., "Midterm CAT", "End Term Exam"
  weight: number; // Percentage, e.g., 30 for 30%
  schoolId: string;
}

export interface Mark {
  studentId: string;
  subjectId: string;
  examSessionId: string;
  score: number | null; // Null if not entered yet
  schoolId: string;
}